matrix 64 bit indexing

+++ src/utils/matrix.h
@@ -19,7 +19,7 @@ namespace utils {
 template <typename T>
 class Matrix {
  public:
-  Matrix(uint32_t n, uint32_t m, T filler = 0)
+  Matrix(uint64_t n, uint64_t m, T filler = 0)
       : n_(n), m_(m), data_(n*m, filler) {}
   Matrix(const Matrix<T>& other)
       : n_(other.n_), m_(other.m_), data_(other.data_) {}
@@ -31,17 +31,17 @@ class Matrix {
     return *this;
   }
 
-  const T& Access(uint32_t i, uint32_t j) const {
+  const T& Access(uint64_t i, uint64_t j) const {
     assert(i < n_);
     assert(j < m_);
     return data_[i*m_ + j];
   }
-  T& operator()(uint32_t i, uint32_t j) {
+  T& operator()(uint64_t i, uint64_t j) {
     return const_cast<T&>(const_cast<const Matrix<T>*>(this)->Access(i, j));
   }
-  const T& operator()(uint32_t i, uint32_t j) const { return Access(i, j); }
+  const T& operator()(uint64_t i, uint64_t j) const { return Access(i, j); }
 
-  VectorView<T> operator()(uint32_t i) const {
+  VectorView<T> operator()(uint64_t i) const {
     assert(i < n_);
     return VectorView<T>(&data_[i*m_], m_);
   }
@@ -51,14 +51,14 @@ class Matrix {
   }
 
   template <typename U>
-  void SetRow(uint32_t i, VectorView<U> r) {
+  void SetRow(uint64_t i, VectorView<U> r) {
     assert(i < n_);
     assert(r.n() == m_);
     std::copy(r.begin(), r.end(), &(*this)(i, 0));
   }
 
   template <typename U>
-  void AddToRow(uint32_t i, VectorView<U> r) {
+  void AddToRow(uint64_t i, VectorView<U> r) {
     assert(i < n_);
     assert(r.n() == m_);
     T* row_begin = &(*this)(i, 0);
@@ -67,7 +67,7 @@ class Matrix {
   }
 
   template <typename U>
-  void SubtractFromRow(uint32_t i, VectorView<U> r) {
+  void SubtractFromRow(uint64_t i, VectorView<U> r) {
     assert(i < n_);
     assert(r.n() == m_);
     T* row_begin = &(*this)(i, 0);
@@ -78,7 +78,7 @@ class Matrix {
   template <typename U>
   void Divide(VectorView<U> a) {
     assert(a.n() == n_);
-    for (uint32_t i = 0; i < n_; ++i) {
+    for (uint64_t i = 0; i < n_; ++i) {
       T* row_begin = &(*this)(i, 0);
       U divisor = a(i);
       assert(divisor != 0);
@@ -92,8 +92,8 @@ class Matrix {
     std::fill(data_.begin(), data_.end(), filler);
   }
 
-  uint32_t n() const { return n_; }
-  uint32_t m() const { return m_; }
+  uint64_t n() const { return n_; }
+  uint64_t m() const { return m_; }
 
   template <class Archive>
   void serialize(Archive& ar) {  // NOLINT(runtime/references)
@@ -101,9 +101,9 @@ class Matrix {
   }
 
   friend std::ostream& operator<<(std::ostream& os, const Matrix<T>& mx) {
-    for (uint32_t i = 0; i < mx.n_; ++i) {
+    for (uint64_t i = 0; i < mx.n_; ++i) {
       os << mx(i, 0);
-      for (uint32_t j = 1; j < mx.m_; ++j) {
+      for (uint64_t j = 1; j < mx.m_; ++j) {
         os << "," << mx(i, j);
       }
       os << std::endl;
@@ -124,17 +124,17 @@ class Matrix {
   data_iter_diff pos(const_data_iter it) const { return it - begin(); }
   data_iter_diff pos(const T* it) const { return it - data_.data(); }
 
-  uint32_t row(const_data_iter it) const { return pos(it) / m_; }
-  uint32_t row(const T* it) const { return pos(it) / m_; }
-  uint32_t row(data_iter_diff pos) const { return pos / m_; }
+  uint64_t row(const_data_iter it) const { return pos(it) / m_; }
+  uint64_t row(const T* it) const { return pos(it) / m_; }
+  uint64_t row(data_iter_diff pos) const { return pos / m_; }
 
-  uint32_t col(const_data_iter it) const { return pos(it) % m_; }
-  uint32_t col(const T* it) const { return pos(it) % m_; }
-  uint32_t col(data_iter_diff pos) const { return pos % m_; }
+  uint64_t col(const_data_iter it) const { return pos(it) % m_; }
+  uint64_t col(const T* it) const { return pos(it) % m_; }
+  uint64_t col(data_iter_diff pos) const { return pos % m_; }
 
  private:
-  uint32_t n_;  // rows
-  uint32_t m_;  // cols
+  uint64_t n_;  // rows
+  uint64_t m_;  // cols
   std::vector<T> data_;
 };  // Matrix
